+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for key value pairs
+  [\#86](https://github.com/osiegmar/logback-gelf/issues/86)
+
 ### Changed
 - Upgrade to Java 11 (Premier Support of Java 8 ended in March 2022).